requirements.txt
msal-streamlit-authentication
python-dotenv
streamlit

psycopg2-binary # postgresql

uvicorn[standard]
pydantic
azure-identity
azure-keyvault